[CHORE] main hotfix 최신화 (#124)
* Update application.properties

* Update application.properties

* [CI/CD] Deploy to EC2 (#122)

Hotfix: notifications tinytext 관련 오류 수정

* Fix: iamport restclient 추가

* Hotfix: portone 관련 모듈 최신화

---------

+++ src/main/java/com/gongspot/project/domain/order/dto/OrderRequestDTO.java
@@ -16,7 +16,7 @@ public class OrderRequestDTO {
     private Long userId;
     private Long productId;
     private PayMethod payMethod;
-    private BigDecimal totalPrice;
+    private Long totalPrice;
     private String merchantUid;
 
 }
